+++ modules/apps/object/object-web/src/main/java/com/liferay/object/web/internal/model/ProxyObjectEntry.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.object.web.internal.util;
+package com.liferay.object.web.internal.model;
 
 import com.liferay.object.model.ObjectEntry;
 import com.liferay.object.model.ObjectEntryWrapper;
